--- src/utils/DarCollectionUtils.js
@@ -1,417 +1,0 @@
-import { chain, cloneDeep, concat, filter, find, findIndex, flatMap, flatten, groupBy, includes, isEmpty, isNil, map, toLower } from 'lodash'
-import { Styles } from 'src/libs/theme.js'
-import { formatDate, Notifications } from '../libs/utils'
-import { Collections } from '../libs/ajax/Collections'
-
-export const rpVoteKey = 'RUS Vote'
-
-// Helper function for processDataUseBuckets, essentially organizes votes in a dar's elections by type
-export const processVotesForBucket = (darElections = []) => {
-  const rp = {
-    chairpersonVotes: [],
-    memberVotes: [],
-    finalVotes: [],
-  }
-  const dataAccess = {
-    finalVotes: [],
-    memberVotes: [],
-    chairpersonVotes: [],
-    agreementVotes: [],
-    radarVotes: [],
-  }
-  darElections.forEach((election) => {
-    const { electionType, votes, status = [] } = election
-    // add field to each vote object to indicate election status
-    const updatedVotes = Object.values(votes).map(vote => ({
-      ...vote,
-      electionStatus: status,
-    }))
-    updatedVotes.forEach((vote) => {
-      votes[vote.voteId] = vote
-    })
-    const dateSortedVotes = updatedVotes.toSorted(vote => vote.updateDate)
-    let targetFinal, targetChair, targetMember, targetFinalType, targetRadar
-
-    if (electionType === 'RP') {
-      targetFinalType = 'chairperson'
-      targetMember = rp.memberVotes
-      targetChair = rp.chairpersonVotes
-      targetFinal = rp.finalVotes
-    }
-    else {
-      targetFinalType = 'final'
-      targetMember = dataAccess.memberVotes
-      targetChair = dataAccess.chairpersonVotes
-      targetFinal = dataAccess.finalVotes
-      targetRadar = dataAccess.radarVotes
-    }
-    dateSortedVotes.forEach((vote) => {
-      const lowerCaseType = toLower(vote.type)
-      switch (lowerCaseType) {
-        case 'radar_approve':
-          targetRadar.push(vote)
-          break
-        case 'chairperson':
-          targetChair.push(vote)
-          break
-        case 'dac':
-          targetMember.push(vote)
-          break
-        default:
-          break
-      }
-      if (lowerCaseType === targetFinalType) {
-        targetFinal.push(vote)
-      }
-    })
-  })
-  return { rp, dataAccess }
-}
-
-// Gets data access votes from this bucket by members of this user's DAC
-// Note that filtering by DAC does not occur if user is viewing on admin review page
-export const extractDacDataAccessVotesFromBucket = (bucket, user, adminPage) => {
-  const votes = bucket?.votes ?? []
-
-  let memberVotesArrays = chain(votes)
-    .map(voteData => voteData.dataAccess)
-    .filter(dataAccessData => !isEmpty(dataAccessData))
-    .map(filteredData => filteredData.memberVotes)
-    .value()
-
-  if (!adminPage) {
-    memberVotesArrays = filterVoteArraysForUsersDac(memberVotesArrays, user)
-  }
-  return flatten(memberVotesArrays)
-}
-
-// Gets rp votes from this bucket by members of this user's DAC
-// Note that filtering by DAC does not occur for users viewing through admin review page
-export const extractDacRPVotesFromBucket = (bucket, user, adminPage) => {
-  const votes = !isNil(bucket) ? bucket.votes : []
-  let rpVoteArrays = chain(votes)
-    .map(voteData => voteData.rp)
-    .filter(rpData => !isEmpty(rpData))
-    .map(filteredData => filteredData.memberVotes)
-    .value()
-
-  if (!adminPage) {
-    rpVoteArrays = filterVoteArraysForUsersDac(rpVoteArrays, user)
-  }
-  return flatten(rpVoteArrays)
-}
-
-// Applies filter to arrays of votes grouped by election and
-// only keeps arrays where at least one vote has the userId of the provided user
-const filterVoteArraysForUsersDac = (voteArrays = [], user) => {
-  const userIdsOfVotes = (votes) => {
-    return map(votes, vote => vote.userId)
-  }
-
-  return filter(voteArrays, voteArray => includes(userIdsOfVotes(voteArray), user.userId))
-}
-
-// Gets this user's data access votes from this bucket; radar, final and chairperson votes if isChair is true, member votes if false
-// Note that filtering by DAC does not occur for users viewing through admin review page
-export const extractUserDataAccessVotesFromBucket = (bucket, user, isChair = false, adminPage = false) => {
-  const votes = !isNil(bucket) ? bucket.votes : []
-  const adminOrChair = adminPage || isChair
-  const userDataAccessVotes = votes.map((voteGroup) => {
-    // If admin page or chair, we want to include all final, chair, and radar votes
-    if (adminOrChair) {
-      const chairpersonVotes = voteGroup.dataAccess?.chairpersonVotes || []
-      const finalVotes = voteGroup.dataAccess?.finalVotes || []
-      const radarVotes = voteGroup.dataAccess?.radarVotes || []
-      return chairpersonVotes.concat(finalVotes, radarVotes)
-    }
-    else {
-      return voteGroup.dataAccess?.memberVotes || []
-    }
-  }).flat(Infinity)
-  if (adminPage) {
-    // If admin page, we want to include all votes regardless of userId
-    return userDataAccessVotes
-  }
-  else {
-    return userDataAccessVotes.filter(vote => vote.userId === user.userId)
-  }
-}
-
-// Gets this user's rp votes from this bucket; chairperson votes if isChair is true, member votes if false
-// Note that filtering by DAC does not occur when viewing through the admin review page
-export const extractUserRPVotesFromBucket = (bucket, user, isChair = false, adminPage = false) => {
-  const votes = !isNil(bucket) ? bucket.votes : []
-  const adminOrChair = adminPage || isChair
-  const userRPVotes = votes?.map((voteGroup) => {
-    if (adminOrChair) {
-      return voteGroup.rp?.chairpersonVotes || []
-    }
-    else {
-      return voteGroup.rp?.memberVotes || []
-    }
-  }).flat(Infinity)
-  if (adminPage) {
-    return userRPVotes?.filter(vote => !isNil(vote.vote)) || []
-  }
-  else {
-    return userRPVotes?.filter(vote => vote.userId === user.userId) || []
-  }
-}
-
-// collapses votes by the same user with same vote (true/false) into a singular vote with appended rationales / dates if different
-export const collapseVotesByUser = (votes) => {
-  const votesGroupedByUser = groupBy(cloneDeep(votes), vote => vote.userId)
-  return flatMap(Object.keys(votesGroupedByUser), (userIdKey) => {
-    const votesByUser = votesGroupedByUser[userIdKey]
-    const collapsedVotes = collapseVotes({ votes: votesByUser })
-    return convertToVoteObjects({ collapsedVotes })
-  })
-}
-
-// helper method to collapse votes by converting them to an object with differing rationales and dates in arrays
-const collapseVotes = ({ votes }) => {
-  const collapsedVotes = {}
-  votes.forEach((vote) => {
-    const matchingVote = collapsedVotes[`${vote.vote}`]
-    const lastUpdate = vote.updateDate
-    if (isNil(matchingVote)) {
-      collapsedVotes[`${vote.vote}`] = {
-        userId: vote.userId,
-        vote: vote.vote,
-        voteId: vote.voteId,
-        displayName: vote.displayName,
-        rationales: !isNil(vote.rationale) ? [vote.rationale] : [],
-        lastUpdates: !isNil(lastUpdate) ? [lastUpdate] : [],
-      }
-    }
-    else {
-      addIfUnique(vote.rationale, matchingVote.rationales)
-      addIfUnique(lastUpdate, matchingVote.lastUpdates)
-    }
-  })
-  return collapsedVotes
-}
-
-// helper method to follow collapseVotes in flow
-const convertToVoteObjects = ({ collapsedVotes }) => {
-  return map(Object.keys(collapsedVotes), (key) => {
-    const collapsedVote = collapsedVotes[key]
-    const collapsedRationale = appendAll(collapsedVote.rationales)
-    const collapsedDate = appendAll(map(collapsedVote.lastUpdates, date => formatDate(date)))
-
-    return {
-      userId: collapsedVote.userId,
-      vote: collapsedVote.vote,
-      voteId: collapsedVote.voteId,
-      displayName: collapsedVote.displayName,
-      rationale: collapsedRationale,
-      lastUpdated: collapsedDate,
-    }
-  })
-}
-
-const appendAll = (values) => {
-  let result = ''
-  values.forEach((value) => {
-    result += `${value}\n`
-  })
-  return !isEmpty(result) ? result : null
-}
-
-const addIfUnique = (newValue, existingValues) => {
-  if (!isNil(newValue) && !includes(existingValues, newValue)) {
-    existingValues.push(newValue)
-  }
-}
-
-export const updateCollectionFn = ({ collections, filterFn, searchText, setCollections, setFilteredList }) =>
-  (updatedCollection) => {
-    const targetIndex = findIndex(collections,
-      collection =>
-        collection.darCollectionId === updatedCollection.darCollectionId,
-    )
-    if (targetIndex < 0) {
-      Notifications.showError({
-        text: `Error: Could not find ${updatedCollection.darCode} collection`,
-      })
-    }
-    else {
-      const collectionsCopy = cloneDeep(collections)
-      collectionsCopy[targetIndex] = updatedCollection
-      const updatedFilteredList = filterFn(
-        searchText,
-        collectionsCopy,
-      )
-      setCollections(collectionsCopy)
-      setFilteredList(updatedFilteredList)
-    }
-  }
-
-export const cancelCollectionFn
-  = ({ updateCollections, role }) =>
-    async ({ darCode, darCollectionId }) => {
-      try {
-        await Collections.cancelCollection(darCollectionId, role)
-        const summary = await Collections.getCollectionSummaryByRoleNameAndId({
-          id: darCollectionId,
-          roleName: role,
-        })
-        updateCollections(summary)
-        Notifications.showSuccess({ text: `Successfully canceled ${darCode}` })
-      }
-      catch (_error) {
-        Notifications.showError({ text: `Error canceling ${darCode}` })
-      }
-    }
-
-export const openCollectionFn
-  = ({ updateCollections, role }) =>
-    async ({ darCode, darCollectionId }) => {
-      try {
-        await Collections.openElectionsById(darCollectionId)
-        const summary = await Collections.getCollectionSummaryByRoleNameAndId({
-          id: darCollectionId,
-          roleName: role,
-        })
-        updateCollections(summary)
-        Notifications.showSuccess({ text: `Successfully opened ${darCode}` })
-      }
-      catch (_error) {
-        Notifications.showError({ text: `Error opening ${darCode}` })
-      }
-    }
-
-export const approveCollectionFn
-  = ({ updateCollections, role }) =>
-    async ({ darCode, darCollectionId }) => {
-      try {
-        await Collections.approveCollectionById(darCollectionId)
-        const summary = await Collections.getCollectionSummaryByRoleNameAndId({
-          id: darCollectionId,
-          roleName: role,
-        })
-        updateCollections(summary)
-        Notifications.showSuccess({ text: `Successfully approved ${darCode}` })
-      }
-      catch (_error) {
-        Notifications.showError({ text: `Error approving ${darCode}` })
-      }
-    }
-
-// helper function used in DarCollectionReview to update final vote on source of truth
-// done to trigger re-renders on parent and child components (vote summary bar, member tab, etc.)
-export const updateFinalVote = ({ key, votePayload, voteIds, dataUseBuckets, setDataUseBuckets }) => {
-  if (!isEmpty(votePayload)) {
-    // clone entire bucket to trigger page re-render on bucket update (setDataUseBuckets)
-    const clonedBuckets = cloneDeep(dataUseBuckets)
-    const isRPBucket = toLower(key) === toLower(rpVoteKey)
-    const targetBucket = find(clonedBuckets, bucket => toLower(bucket.key) === toLower(key))
-    // source of votes will differ depending on the bucket (rp vs non-rp), so determine the callback function for flow here
-    const voteObjectCallback = isRPBucket ? voteObj => voteObj.rp : voteObj => voteObj.dataAccess
-    // to keep local source of truth updated without a fetch, we will need to update both the final and the chairperson votes
-    // to make searching on the votes easier, concatenate and then flatten the finalVotes and chairpersonVotes into one array
-    // NOTE: For the RP bucket the chairperson votes and the final votes are the same (RP has no final vote)
-    // This was a conscious choice in order to keep processing the same between RP and non-RP buckets
-    const votes = chain(targetBucket.votes)
-      .map(voteObjectCallback)
-      .flatMap(voteObj => concat(voteObj.finalVotes, voteObj.chairpersonVotes))
-      .value()
-
-    // perform in place update of vote and vote rationale based on voteIds arguments
-    // updates to the vote here will be reflected in clonedBuckets since the vote references are the same
-    chain(votes)
-      .filter(vote => includes(voteIds, vote.voteId))
-      .forEach((currentVote) => {
-        const { rationale, vote } = votePayload
-        currentVote.rationale = rationale
-        currentVote.vote = vote
-      })
-      .value()
-    // set new bucket to trigger re-render, return clonedBuckets for debugging/testing efforts
-    setDataUseBuckets(clonedBuckets)
-    return clonedBuckets
-  }
-}
-
-export const consoleTypes = {
-  ADMIN: 'admin',
-  MEMBER: 'member',
-  MANAGE_ACCESS: 'manageAccess',
-  CHAIR: 'chair',
-  SIGNING_OFFICIAL: 'signingOfficial',
-  RESEARCHER: 'researcher',
-}
-
-export const styles = {
-  baseStyle: {
-    fontFamily: 'Montserrat',
-    fontSize: '1.6rem',
-    fontWeight: 400,
-    display: 'flex',
-    padding: '1rem 2%',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    whiteSpace: 'pre-wrap',
-    backgroundColor: 'white',
-    border: '1px solid #DEDEDE',
-    margin: '0.5% 0',
-  },
-  columnStyle: {
-    ...Styles.TABLE.HEADER_ROW, justifyContent: 'space-between',
-    color: '#7B7B7B',
-    fontFamily: 'Montserrat',
-    fontSize: '1.2rem',
-    fontWeight: 'bold',
-    letterSpacing: '0.2px',
-    textTransform: 'uppercase',
-    backgroundColor: 'B8CDD3',
-    border: 'none',
-  },
-  cellWidth: {
-    darCode: '10%',
-    dacNames: '8%',
-    projectTitle: '13%',
-    submissionDate: '8%',
-    researcher: '9%',
-    institution: '11.5%',
-    datasetCount: '6%',
-    expirationDate: '8%',
-    status: '9%',
-    actions: '13%',
-  },
-  color: {
-    darCode: '#000000',
-    dacNames: '#000000',
-    projectTitle: '#000000',
-    submissionDate: '#000000',
-    researcher: '#000000',
-    institution: '#354052',
-    datasetCount: '#354052',
-    status: '#000000',
-    actions: '#000000',
-  },
-  fontSize: {
-    darCode: '1.6rem',
-    dacNames: '1.4rem',
-    projectTitle: '1.4rem',
-    submissionDate: '1.4rem',
-    researcher: '1.4rem',
-    institution: '1.4rem',
-    datasetCount: '2.0rem',
-    status: '1.6rem',
-    actions: '1.6rem',
-  },
-}
-
-export const DarCollectionTableColumnOptions = {
-  DAR_CODE: 'darCode',
-  DAC: 'dacNames',
-  NAME: 'name',
-  SUBMISSION_DATE: 'submissionDate',
-  RESEARCHER: 'researcher',
-  INSTITUTION: 'institution',
-  DATASET_COUNT: 'datasetCount',
-  EXPIRES_AT: 'expiresAt',
-  STATUS: 'status',
-  ACTIONS: 'actions',
-}
